database has set

+++ app/src/main/java/com/example/moneyeverydayxml/history/data/TransactionEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "transactions")
 data class TransactionEntity(
-    @PrimaryKey(autoGenerate = true) var id: Int? = null,
+    @PrimaryKey(autoGenerate = true) var id: Long? = null,
     @ColumnInfo(name = "time") val time: Long,
     @ColumnInfo(name = "date") val date: String,
     @ColumnInfo(name = "count") val count: String
